requirements.txt
APScheduler==3.6.3
blinker==1.6.2
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.5.0
pymongo==4.3.3
python-telegram-bot==13.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.29.0
six==1.16.0
tornado==6.3.1
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
Werkzeug==2.3.3
base58==2.1.1  # Added for Solana address validation
